clock: revert most of "STOPWATCH first Parts"
Named commit breaks things even if STOP_WATCH isn't even defined.
I tried to keep the good parts, but had to remove most of the
hindsight all this conditional compilation stuff seems to be
unmaintainable in the long run. We should refactor the clock
code to fit in alternative clock modes more easily.